[REFACTORY] Refatoração de código com objetivo de componentizar as tabelas.
